wrdsConnect {wrdsr}	R Documentation

Establish a connection to WRDS in R

Description

wrdsConnect returns a connection to the WRDS database

Usage

wrdsConnect(username, password, librefs = NULL, conn = "postgresql")

Arguments

username
Your WRDS username as a character string

password
Your WRDS password as a character string

librefs
An optional paramater that can be used to assign SAS libraries. This should be a character string of the form "library_name 'location'". If you have multiple libraries then they should be seperated by a semicolon like this, "library_name1 location1';library_name2 'location2';library_name3 'location3". Unfortunately, it does not seem to be possible to assign multiple locations to a single library using a JDBC connection, so you will have to provide unique library names for each location. Most of the WRDS data is already preassigned to a library, so you shouldn't need to assign library names very often anyway.

conn
Specify whether you want to connect using 'postgresql' or 'sas'

Value

wrdsConnect will return a DBIConnection object that can be used with DBI functions such as dbSendQuery to access the WRDS database.

Examples

## Not run: 
wrds <- wrdsConnect("your_wrds_username", "your_wrds_password")
sql <- "SELECT conm, fyear, at
        FROM compm.funda
        WHERE fyear >= 2010"
res <- dbSendQuery(wrds, sql)
data <- dbFetch(res)

wrds <- wrdsConnect("jestover", "Aetnaus758", "SEC '/wrds/sec/sasdata'")
sql <- "SELECT *
        FROM sec.wforms2"
res <- dbSendQuery(wrds, sql)
data <- dbFetch(res)

## End(Not run)

[Package wrdsr version 0.1.0 Index]